-Добавлен вывод убойки как на контроллере(экспериментальный) -Команда tsmp_weapon_log_hits, значения 0/1 включает/выключает вывод убойки в лог, по умолчанию 0. -Listplayers теперь имеет формат как у контроллера, оригинальный вызывается по команде sv_listplayers_original -Поправлены некоторые предупреждения -Отключен вывод ненужной информации в лог -XrLc: вернул скрытый поток просчета освещения mu

+++ xrCore/_std_extensions.h
@@ -29,7 +29,7 @@
 #undef max
 #endif
 
-IC u32 xr_strlen(const char* S) { return strlen(S); }
+IC u32 xr_strlen(const char* S) { return (u32)strlen(S); }
 //#define strlen(a) xr_strlen(a)
 
 #ifdef  _EDITOR
